demo: show timetable info in drawer
